dcerpc: asm.h is no longer available on newer Linux kernels
Due to the way the PAGE_SIZE macro is used, it is not possible
to use sysconf() on Linux to get the actual page size.  So default
to 4096 which is at least what we had before and the usual default.

(r30217)
(cherry picked from commit d6d98630fc0889911ff51afa4280b7e5f4a3c774)