API: Add rpc_server_use_protseq_socket API for launchd support.
This change adds a new public API, rpc_server_use_protseq_socket().
this API is used to take a socket represented with a platform-specific
data structure, and insert it into the listening set of the appropriate
network handler. This allows servers to take file descriptors vended
by launchd and pump them down into the RPC runtime.

To this end, we add a new EPV operation to the socket layer to
duplicate a native socket representation to a rpc_socket_t, and
another operation to the network layer to add an existinc rpc_socket_t
to the listen set. We then plumb these operations through for BSD
socket types and the ncacn network type.