build: Make buildit builds work for i386 and x86_64.
We need a number of tweaks in the outer-level build system to make
sure we get the CFLAGS we want and that we don't engage autoconf's
cross-compilation support. See README for a longer explanation.

We also do some cleanup of $DSTROOT so that we don't install demo
files, libtool cruft and unused headers.

Finally, we copy biinaries to $SYMROOT so the builder can generate
dSYMs.