idl: Support an allocator context pointer.
This change propagates an allocator context pointer throughout the
runtime.

This isn't used on the server side, since the server manages memory
with a mem_handle. On the client side, this allows the caller to
plug in a custom memory pool to track all the memory allocated for
storing the [out] parameters.

This is a source and binary incompatible change. Any code that uses
the rpc_ss_swap_client_alloc_free() and set of APIs will need to
be changed.