DCERPC: Rename uuid_t to idl_uuid_t.
The uuid_t type conflicts with Darwin's uuid_t type even though
they are both the same under the covers. Likewise hacked around
this with the preprocessor but doing this creates awful issues for
code that actually needs to use the Darwin definition.

The solution is for DCE/RPC to use idl_uuid_t and rename all the
many usages.