IDL compiler: Generate warning-free stub code.
Using {0} to initialize arbitrary structures and unions causes GCC
to emit warnings when -Wmissing-field-initializers and -Wmissing-braces
are used.  However the functionally equivalent {} does not result
in any warnings.

We also put in place a couple of general hooks to suppress warnings
in the generated stub code (assuming the compiler implements the
suppressions correctly, unlike gcc).