idl_compiler: User re-entrant parsers and scanners.
Depend on modern Flex and Bison to generate pure (re-entrant) IDL
and ACF parsers and lexers. For both the NIDL and the ACF parsers,
we wrap the parser and scanner up in a parser object and publish a
small external API to drive it. Instead of using globals to access
the current parser position, we fetch the parser_location_t from
the current parser and pass that around as needed. Unfortunately
the previous reliance on globals means that some parts of the
compiler infrastructure can emit parser errors when we are not
actually parsing (ie. there is no current parser location).