build: Use Xcode Copy Headers instead of copy files.
If we turn DerivedSources/dce into a folder reference rather than
a group reference, then we can get Xcode to create the subdirectory
in Headers. Then when we do a Copy Files for the rest of the headers,
it is already created.

This means that we don't need the copy-dce-headers script that was
relying on Xcode variables whose values change between SDKs.

Radar: <rdar://problem/7917316>