build: Use Xcode's Copy Files phase to install headers
It seemed like we could use Copy Headers, but buildit tries to copy
the headers before they are generated. If we try ti install the
static DCE headers, then Xcode won't let us compile the IDL files
in include/dce.

The only solution I can see is the one Brad uses, which is to install
a junk header to create the Headers directory (and optional symlink).
The use Copy Files to do the installation, and post-process the
framework to remove the junk. Sigh.

Radar: <rdar://problem/7917316>