Fix the macOS build.
New versions of Xcode seem to pay closer attention to target
dependencies, so we can no longer compile `ep.idl` in both the
"dce" and "nck_stubs" targets. It is a bit hacky, but we still
need to compile this in "nck_stubs" to get the symbols for the
final framework build.

Remove CrashReporter integration. This used an internal Apple
interface that I don't think anyone really needs any more.

Based on a patch by Tim Perfitt <tperfitt@twocanoes.com>.