perf: build perf test tools
Make the perf test tools build. This mostly involved updating code to
use newer style macros, and removing ancient poerability cruft.

The main unsolved problem is that there are tests that depend on DCE
thread cancellatino semantics that don't exist in with POSIX threads.
This is one of the (probably numerous) issues that prevents the tests
from actually working.

Signed-off-by: James Peach <jpeach@apache.org>