ci: add a macOS build CI action
Add a macOS build job to build the Release configuration..

Fix the autoconf script to pass through unknown CPU architectures. The
GitHub macOS builder ends up with an architecture of "i486", which is
weird, but it only matters that we have something.

Set the macOS deplopyment target. Otherwise, we end up building a
version of dceidl that can't acrually run on the host OS version.

Signed-off-by: James Peach <jpeach@apache.org>